Patch: code refactoring (autoloading related)
