+++ init.rb
@@ -1,5 +1,5 @@
-require_dependency File.dirname(__FILE__) + "/lib/issue_view_columns/project_helper_patch.rb"
-require_dependency File.dirname(__FILE__) + "/lib/issue_view_columns/view_issues_show_hook.rb"
+require_dependency File.dirname(__FILE__) + "/lib/redmine_issue_view_columns/project_helper_patch.rb"
+require_dependency File.dirname(__FILE__) + "/lib/redmine_issue_view_columns/view_issues_show_hook.rb"
 
 Redmine::Plugin.register :redmine_issue_view_columns do
   name "Redmine Issue View Columns"
@@ -14,9 +14,6 @@ Redmine::Plugin.register :redmine_issue_view_columns do
   settings default: { "empty": true }, partial: "settings/issue_view_columns_settings"
 end
 
-if Rails.configuration.respond_to?(:autoloader) && Rails.configuration.autoloader == :zeitwerk
-  Rails.autoloaders.each { |loader| loader.ignore(File.dirname(__FILE__) + '/lib') }
-end
 
 # helper methods needed for the Settings page of the project also
 ProjectsController.send :helper, IssueViewColumnsHelper
