+++ lib/redmine_issue_view_columns/project_helper_patch.rb
@@ -1,0 +1,17 @@
+module RedmineIssueViewColumns
+  module ProjectHelperPatch
+    def project_settings_tabs
+      super.tap do |tabs|
+        tabs << {
+          name: "issue_view_columns",
+          action: :issue_view_columns,
+          partial: "issue_view_columns/index",
+          label: :issue_view_columns_settings,
+        } if User.current.allowed_to?(:manage_issue_view_columns, @project) &&
+             @project.module_enabled?(:issue_view_columns)
+      end
+    end
+  end
+end
+
+ProjectsController.helper RedmineIssueViewColumns::ProjectHelperPatch
