+++ lib/redmine_issue_view_columns/view_issues_show_hook.rb
@@ -1,0 +1,5 @@
+module RedmineIssueViewColumns
+  class ViewIssuesShowHook < Redmine::Hook::ViewListener
+    render_on :view_issues_show_description_bottom, partial: 'issue_view_columns/custom_subtasks'
+  end
+end
